recipes: switchdev: span: Remove tc-matchall parsing
As for the kernel commit 7a335adad8b0 ("net/sched: cls_matchall: Dump the
classifier flags"), the output format of the "tc filter show" command has
changed to includes the flags identification. Due to the fact that the
switchdev-span recipe parses the tc output, it failed to parse correctly
the output of that command.

Remove the tc output parsing as it can be avoided. This makes the test
more robust and not depend on the specific tc output format.

Signed-off-by: Yotam Gigi <yotamg@mellanox.com>
Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>