recipes: switchdev: span: Add ipv6 support
Before, the test failed when ipv6 was turned on due to the fact that the
tcpdump filter matched only on the interfaces ipv4 addresses and only on
ipv4 icmp packets.

Make the test run different tcpdump instances for ipv4 and ipv6 separately
and fix the tcpdump filter to match on the correct protocol and addresses.
In addition, ipv6 icmp packets include neighbour advertisement and
neighbour solicitation control packets, which are sent quite often, so
enhance the tcpdump filter to match only echo-requests and echo-replies.

Signed-off-by: Yotam Gigi <yotamg@mellanox.com>
Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>