add lnst.Slave.Job
Defines classes of jobs to be run on a slave. Based on the old
lnst.Common.NetTestCommand code, but simplified to work with newer 'Job'
implementation that started on the Controller.

Defines a Job class that is used as a container object by the Slave to
fork and run a GenericJob object. For now only a ShellExecJob and a
ModuleJob are supported. Previous "control" commands are now only
methods of the Job container object.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>