NetTestSlave: create dynamic lnst.Tests package
The lnst.Tests package is present on the Controller and provides
upstream maintained test classes. These can be instantiated and used to
run tests on the Slaves. While doing so they're sent over the network to
the Slaves where they're dynamically imported. Since the lnst.Tests
package doesn't originaly exist on the slave, every imported module
imported this way will result in a python runtime warning of the package
not existing. This doesn't break anything but it does add clutter to the
logs.
Pre-creating the package on the slave will fix the issue.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>