lnst.Common.Parameters: add DeviceParam
This parameter class accepts a Device or DeviceRef objects as possible
values. Since RemoteDevice is also recognized as a Device instance this
will work on the Controller where the parameter value is set.

DeviceRef is then used by the Controller-Slave communication methods to
transfer the information. For this to work properly the TestModule is
deepcopied and modified (only DeviceParam) before transmission. Since
DeviceParam can refer to a VirtualDevice object it cannot be deepcopied
so we override it's deepcopy operation to only do a shallowcopy.

It's important to note that at the moment this works in just one
direction: Controller -> Slave and not the other way around. However
this shouldn't be a problem since passing Parameters the other way
around doesn't really make sense.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>