recipes: switchdev: TestLib: Make the pktgen dst_eth controllable
Before this commit, the Testlib.pktgen packet's destination MAC address was
determined by the MAC address of the destination port. That cannot be used
when L3 routing is needed, as the destination MAC should be the MAC address
of the gateway.

Make the TestLib.pktgen function more flexible by allowing the caller to
determine the dst_eth.

Signed-off-by: Yotam Gigi <yotamg@mellanox.com>
Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>