recipes: switchdev: Add recipe for ECMP with link down
Test the behaviour of ECMP topology when some of the ECMP links go down,
either with ignore_routes_with_linkdown sysctl option is set and not.

When ignore_routes_with_linkdown is set, the traffic should not be affected
at all and there should not be any packet-loss due to ECMP links going down.
When ignore_routes_with_linkdown is not set, expect traffic loss on the
ECMP connection.

Signed-off-by: Yotam Gigi <yotamg@mellanox.com>
Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>