recipes/tunnels: add VXLAN/Geneve tests
This patch adds new test for testing VXLAN or Geneve tunnels. The test
scenario consists of two baremetal hosts, each with one NIC. On both
hosts, there is an OVS bridge, where number of internal devices are
created along with the same number of VXLAN/Geneve interfaces. The
number is set in the task in TUNNEL_COUNT variable. This variable is set
to 16 by default. Netperf is run in parallel over each tunnel and over each
tunnel number of STREAM are run in parallel in order to maximize link
utilization. This number is specified with nperf_num_parallel alias and is set
to 4 by default.

The result throughputs are then summarized and displayed with Custom module.
This is temporary solution before we implement more complex parallel netperf in
LNST test module.

Signed-off-by: Jiri Prochazka <jprochaz@redhat.com>
Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>