PerfRepoUtils: add support for RR test in perfrepo templates
This patch adds test_type param to all PerfRepo template functions. When
this param is not specified, it is expected the test to be *_STREAM and
metric labels are throughput, throughput_deviation, etc. When the param
is set to RR, metric labels used are rr_rate, rr_rate_deviation, etc.

This is needed for the new short-lived connections test, where the unit
we measure is not throughput but request/response rate.

Signed-off-by: Jiri Prochazka <jprochaz@redhat.com>
Acked-by: Jan Tluka <jtluka@redhat.com>