test_modules/Netperf: make test recognition prettier
In _parse_threshold() the recognition of test type (STREAM and RR) is
done via long if statement, when we only need to check for STREAM or RR
in test type name. This also fixes TCP_CRR test threshold parsing as
that test type was not present in the if condition.

Signed-off-by: Jiri Prochazka <jprochaz@redhat.com>
Acked-by: Jan Tluka <jtluka@redhat.com>