test_modules/Netperf: fix output format in execution of STREAM tests
In current version of netperf (2.7.1) there is a bug with formatting
option b, which should represent bits, but it represents Bytes instead
(the same as option B). Due to this bug, we are reverting back to
kilobits.

Thanks to Jan Tluka (jtluka@redhat.com) for discovering this issue.

Fixes: a7f6aad8 ("test_modules/netperf: use bit/s in STREAM tests and
parse RR tests")

Signed-off-by: Jiri Prochazka <jprochaz@redhat.com>
Acked-by: Jan Tluka <jtluka@redhat.com>