add lnst.Controller.Namespace
Defines the abstract class Namespace that represents the base of both the
Host (root namespace) and NetNamespace (other network namespaces) classes. It
defines the tester facing APIs which should be mostly identical for both, with
slight differences. The Namespace class should never be instantiated on it's
own, only the derived classes should be created.

This is the beggining of adding namespaces support to Python recipes.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>