Host: rewrite to inherit from Namespace
Redefines the Host class to be inherited from the base Namespace class.
The Host class now represents the root namespace of a slave machine. In
addition to the base Namespace class it implements the 'namespaces'
property which returns the list of other network namespaces and it also
allows for a NetNamespace object to be assigned to the Host object which
will in turn create the network namespace on the slave.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>