Utils: allow non-secure md5 use
This will fix the issue with LNST on FIPS enabled system. FIPS does not allow
use of md5. Python's hashlib has a parameter that can be used to allow use of
md5 for non-security purpose. LNST uses md5 for caching test modules and tools
only.

Tested both with python2.6 and python2.7

Signed-off-by: Jan Tluka <jtluka@redhat.com>
Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>