SlavePool: fix unresolvable hostname exception
If LNST controller processes a pool that contains a machine with a hostname
that can't be resolved by DNS the controller fails with unhandled exception.
The exception is raised on shutdown of the socket used to connect to the
machine. Since the connection on the socket did not succeed the attempt to
shutdown the connection is an error. Unfortunately this error is not detected
by checking SO_ERR in getsockopt() after calling select() on the socket.

The fix is to mark the machine unavailable immediately unless the error code
raised by connect call is EINPROGRESS meaning the connection proceeded to
establishing phase (the hostname was succesfully resolved).

Fixes: 5498ba2 ("SlavePool: properly close sockets after connection check")
Fixes issue #188

Signed-off-by: Jan Tluka <jtluka@redhat.com>
Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>