lnst.Common.ConnectionHandler: add timeout to _check_connections
This adds a timeout parameter to the _check_connections method of the
ConnectionHandler. This is used in the select call to limit the sleep
time.

The NetTestSlave module defines a MAX_SERVER_HANG constant with value 5,
which is used by the Slave server for the ConnectionHandler timeout.
The main reason for this is to remove the netlink socket from the Server
and make it self managed by the InterfaceManager and we want a quick
turnaround in the Server loop.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>