lnst.Devices.Device: add __setattr__ and modify enable/disable logic
Adding __setattr__ method to recognize when the Device is in a disabled
state and throw an exception when trying to modify anything.

The enabled/disabled logic should only apply to the public API of the
Device derived classes - access to methods/attributes starting with '_'
should be unaffected. When the Device is disabled access to public
methods or attempts to assign to public properties should raise the
DeviceDisabled exception.

I also adjusted the __getattribute__ method to also work this way.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>