InterfaceManager: disable new Device objects by default
When a new Device object is created in reaction to a RTM_NEWLINK
message, it is disabled after it is added to the InterfaceManager
database. This ensures that all Devices we are disabled by default and
you need to explicitly change this to use them.

Soft devices that were created during recipe execution are enabled by
default. Devices mapped by the Controller Mapper class are also
automatically enabled.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>