lnst.Devices.*: update SoftDevice creation implementation
Tested for most of the devices, parameters supplied to the constructor
are translated into Netlink attributes at two levels - the link level
and the link info data level. On the link level the attributes are added
to the attributes of the top level of the netlink message. On the link
info data level, the attributes are added to the nested
IFLA_LINKINFO->IFLA_INFO_DATA attribute list.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>