Interface: add mcast flood and mrouter set functions
Add functions to set mcast flood on\off on an interface that is a bridge
slave. Add a similar support to set mrouter port state.
Mcast flood flag decides whether multicast packets should be flooded to a
port when multicast snooping is disabled. Default is on.
Mrouter port state can be fixed on, fixed off or learning, where the later
indicates that the mrouter state should be learned according to IGMP
queries. If a port is set to be mrouter port, all multicast packets will be
flooded to it, when multicast snooping is enabled.

Signed-off-by: Nogah Frankel <nogahf@mellanox.com>
Reviewed-by: Yotam Gigi <yotamg@mellanox.com>
Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>