recipes: switchdev: TestLib: Check whether packets goes in slowpath
Add a function that validates that the amount of packets went via slowpath
is below a threshold. It counts the traffic with ifstat, meaning that it
returns the traffic count since the last call to it. So one should call it
and ignore the results before starting the traffic.
Add this check to test 20 to check that data is not going via slowpath.

Signed-off-by: Nogah Frankel <nogahf@mellanox.com>
Reviewed-by: Yotam Gigi <yotamg@mellanox.com>
Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>