recipes: switchdev: Add basic mrouter test
This test tests MC flood to ports that is configured statically to be
mrouter, meaning that it should get all the MC traffic, regardless of join
requests.

The test uses 4 ports: one that sends the data, one that send no join
request and shouldn't get any traffic, one that sends some join requests
and should get some of the traffic and a PUT (port under tests).

The test use 3 MC groups:
 - Unregistered one (no port sent join request to it).
 - A MC group that was registered to only by a port that is not the PUT.
 - A MC group that is registered to by the previous port and the PUT.

The test's algorithm is as follows:
- Send the needed join requests.
- Set the PUT to be mrouter (statically on)
- send traffic from all of the MC groups.
 - Check that all of them were sent to the PUT.
 - Check that the other ports got the traffic according to the join
   requests
- Set the PUT to not be mrouter (statically off)
- Send traffic from all of the MC groups.
 - Check that all the ports (including the PUT) got the traffic according
   to the join requests
- Check that all the traffic hadn't gone via slowpath

Signed-off-by: Nogah Frankel <nogahf@mellanox.com>
Reviewed-by: Yotam Gigi <yotamg@mellanox.com>
Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>