Devices.RemoteDevice: deepcopy support
Sometimes we deepcopy objects that contain references to RemoteDevice
objects. However since RemoteDevice objects contains a reference to the
Machine object which can't be deepcopied (PyCapsule objects) this
implements a custom __deepcopy__ method for the RemoteDevice object that
deepcopies the object as normal, but shallowcopies the Machine and
Namespace references.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>