lnst.Common.Parameters: rework of Parameters design
The Parameters class now serves as a simple container for actual
values(objects) of the Parameters while individual Param derived classes
serve as type checkers and optionally hold a default value or boolean
describing if the Parameter is mandatory.

Since Parameters are directly interfaced with different
internal/external parts of LNST, this commit also contains updates to
those.

Parameters for Recipe and/or TestModules are type and mandatory checked
during object initialization. Optionally if the default value is to be
used it is deepcopied from the class Param instance. Parameter values
directly provided by the Tester are not copied at all and the same
instances are used.

For both Recipe and TestModule derived classes, the parameter values can
be accessed like this:

Param definition:
int_param = IntParam(mandatory=True)
optional_param = IntParam()

Param value access in Recipe or Testmodule class:
x = self.params.int_param
if "optional_param" in self.params:
    x += self.params.optional_param

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>