NetTestSlave: Add mcast routing operations
The Linux kernel multicast routing support is done via the mroute socket.
Using setsockopt on that socket, user can:
 - Add/Remove MFCs - MFC (Multicast Forwarding Cache) is practically the
   multicast counterpart to a unicast route.
 - Add/Remove VIFs - VIF (Virtual Interface) is a way to configure a port
   to do multicast forwarding. When a VIF is assigned to a port, the VIF
   index is used when adding an MFC entry.
 - Initialize PIM support.
 - Create the pimreg VIF - A specific interface dedicated for the PIM
   protocol.

In addition, the kernel sends multicast routing notifications to that
socket in a form of IGMP packets.

Implement these operations in the NetTestSlave module. In addition, make
the machine_cleanup function clean the added MFC and VIFs by closing the
mroute socket.

Signed-off-by: Nogah Frankel <nogahf@mellanox.com>
Signed-off-by: Yotam Gigi <yotamg@mellanox.com>
Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>