Machine: Add mcast routing operations
Add the multicast routing operations to the Machine module, by forwarding
the method calls through the RPC interface to the NetTestSlave module.

The VIF add/delete interface is put in the Interface class as it is an
operation done upon an interface. The other operations are added to
Machine class.

Signed-off-by: Nogah Frankel <nogahf@mellanox.com>
Signed-off-by: Yotam Gigi <yotamg@mellanox.com>
Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>