recipes: switchdev: Add common code for multicast router recipes
The next patches in this patch-set are going to introduce several multicast
router tests. Add common topology and common module for these tests.

The module allows to add/remove VIFs and MFCs. In addition, the module can
test multicast forwarding by running traffic and verifying the traffic was
forwarded correctly.

The module follows the kernel API and validates that the added MFC and VIFs
behave as expected. The module tests that:
 - The traffic is forwarded according to the route output interfaces.
 - If there is a pimreg interface in the route, test that the WHOLEPKT
   notifications were sent.
 - When packet is sent to a wrong interface, validate that the matching
   notification is sent.

Signed-off-by: Yotam Gigi <yotamg@mellanox.com>
Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>