recipes: switchdev: Add the multicast router unresolved VIF RIF recipe
The test aims to verify that MFC routes with output interfaces that does
not have a valid RIF, are still functional.

The test adds a simple MFC route and invalidates some of the RIFs by
enslaving the ports to a lag. The test than validates that the route is
still operational.

Signed-off-by: Yotam Gigi <yotamg@mellanox.com>
Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>