recipes: switchdev: Add the multicast router nocache route recipe
The test aims to test the flow of having traffic before the matching route
is created. The traffic creates an "unresolved route" in the kernel, which
holds the no-cache packets for some time interval. If a matching route is
put at that interval, it "resolves" the route and the stored packets are
forwarded according to it.

Test that that flow is functional by triggering traffic for a route before
the actual route is inserted.

The test algorithm is:
 - For 10 random (S,G) routes
    - Trigger traffic that would match the route
    - Test that a NOTIF_NOCACHE mroute notification was received
    - Add the actual route
    - Test that the route is functional

Signed-off-by: Yotam Gigi <yotamg@mellanox.com>
Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>