lnst.Controller.Controller: fix VirtualDevice creation
Since Parameters implementation was changed, accessing undefined
parameters raises an exception instead of returning None. This created a
bug in the map_device code that creates VirtualDevice objects for
virtually matched recipe runs. This fixes the bug by using the getattr
built-in function that returns a default value of None when the
parameter is undefined.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>