lnst.Slave.Job: modify Job exception handling
All exceptions should now be properly logged including tracebacks. The
related exception isn't reraised on the Controller for now because it
can't be caught by the developer when it's a background job.

The Slave.Job class is LNST internal - the message it generates is part
of the internal LNST Ctl-Slave communication protocol.

The Slave.GenericJob class is also LNST internal and separates different
job types (e.g. Shell exec and Test module). Its run method returns data
used by LNST internals to report the results - the "passed" status, the
"type" of the result (e.g. normal result, exception etc...) and holds
additional data (e.g. returned by the test module) that can be used in
the recipe by the tester.

Finally this result data can have any format the developer of the job
type decides on.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>