recipes: switchdev: Test GRE tunnels
This adds a number of elementary test for GRE tunneling in a
configuration where overlay and underlay are each in a different VRF.
Several lifetime scenarios are tested, as well as several features, e.g.
tunnel keys and checksums.

Along with this test is added a new library ipip_common, with common
code used across a number of IPIP scenarios to be added next.

Also added is an XML library with description of a GRE network that is
likewise used by several scenarios in the IPIP suite.

Signed-off-by: Petr Machata <petrm@mellanox.com>
Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>