lnst.Devices: reimplementation of Device options
All options are now properties, with setters where available.

The Device creation implementation now also uses these setters, they've
been reimplemented so that they don't immediately send data to the
kernel, but instead construct a 'bulk' netlink message that will be sent
to the kernel on the _nl_sync call.

All setters directly call the nl_sync method with the operation "set",
but with argument bulk set to False. This way the method is only
executed if bulk mode is enabled for the Device object.

By default all new Device objects are created with bulk mode enabled -
all property setters just update the update message. When the
initialization of the Device object is finished and the bulk message is
sent to the kernel to create the device, the Device object is switched
into non-bulk mode. This means that all setter calls should be
immediately sent to the kernel.

Manipulating the bulk mode of a Device is for now just an LNST internal
fuctionality but later we might expose this to the user/tester.

Calling dir() on a RemoteDevice object (Device wrapper on the
controller) will now forward the dir call on the wrapped *Device object.

Finally I added the _vars method to the Device class that should expose
all properties of the device object, this is experimental but if we like
it we can later expose it to users as well.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>