lnst.Common.Version: modify version detection and comparison
LNSTVersion is now a class instantiated as a module object lnst_version.
This object is then used by both the Slave and Controller for
comparison.

This commit also adds a new module lnst.Common.GitVersion which will
only be present in the git repository (will not be installed by setup.py
and shouldn't be included in packages). This module defines a method
that detects a git instalation and returns the current revision hash
which is appended to the version number. If the module is present, LNST
runs as a development version - failed version comparison only warns
about the different versions. In case something goes wrong during the
git version detection (e.g. git isn't installed) LNST won't start and
return an exception.

If the module isn't present it's assumed that LNST is installed as a
stable version with the version number defined in
lnst.Common.Version.LNSTVersion._version and a failed version check will
report an error and the recipe execution won't continue.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Acked-by: Jan Tluka <jtluka@redhat.com>