lnst.Controller.Recipe: add Controller reference
Instead of just adding the list of hosts to be accessed from the recipe
we add the controller so that we can also call some generic methods,
e.g. waiting for a condition.

This also changes the .matched attribute to a property that accesses the
appropriate attribute of the Controller object.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>