lnst.Controller.MessageDispatcher: wait_for_condition implementation
The new wait_for_condition method is based on the older wait_for_finish,
that passively waited for socket messages and looked for a job_finish
message. The new method instead waits for socket messages for one second
in a loop, on each iteration we call a provided condition method that
MUST return a boolean and stop the wait when the return value is True
(condition matched).

The old wait_for_finish method was renamed to wait_for_job_finish that
now just calls the wait_for_condition method with a condition method
that checks the Job.finished property.

Finally this patch merges the wait_for_result method with the
send_message method, as every rpc message is expected to receive a
result from the remote message call.

Commit also includes some updates to the Machine class replacing calls
the old methods with the new ones.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>