lnst.Controller.Controller: implement wait method
The wait method implements the old "wait for x seconds" functionality
we've had in the old LNST. The method is exported to the tester/recipe
throught the Controller class and uses the newly implemented
wait_for_condition MessageDispatched method.

This implementation is an improvement compared to the older
functionality as the Controller is still receiving and processing
messages from slaves so conditions can wait for a specific update to the
state of a slave and logging will still work during while we're waiting.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>