add lnst.Controller.RecipeResults
This module defines classes for storing Result data related to a test
run. Most are generated automatically by LNST during test execution and
a tester also has a Recipe interface available to create Result objects
for custom entries.

These will later be integrated with the rest of the LNST Controller so
that they're created in the right places and accessible as Recipe data
after the test run is finished for optional post-processing.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>