Recipe: add class RecipeRun
Instances of the RecipeRun class represent a single test run of a
recipe. They store information about Results (executed Jobs or tester
defined) and about the matchine match that was used.

A recipe instance stores a list of these RecipeRun objects and all of
them can be accessed after testing for optional post-processing.

The Recipe class now also has an interface method for the tester which
can be used to add Result objects to the current run, this is intended
to be used from within the tester defined "test" method.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>