lnst.Controller.Controller: use the Recipe instance for test init
When a recipe is succesfully matched and the Controller starts setting
up the Machine objects we should pass the Recipe instance. This is
important since the Machine object needs to be able to generate
JobResult objects.

Another minor improvement is that a correct recipe name will be used
(the name of the Recipe class) instead of the filename of the currently
running executable which was incorrectly used until now.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>