lnst.Controller.Job: refactoring
Removing the netns parameter from the __init__ method and renaming the
host parameter to namespace.

The parameter expects a Namespace instance (a Host or a NetNamespace),
the host can be accessed through a new property "host" and the "netns"
proerty now returns the original Namespace instance the Job was created
with (these two can be identical if the Job is running on the initial
namespace of a slave).

I also moved the id property higher since I think it makes a bit more
sense to have it at the beginning of the class definition....

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>