recipes: switchdev: ipip: Test TTL of a IP-in-IP tunnel
From: Petr Machata <petrm@mellanox.com>

This test file keeps tests related to TTL setting of IP-in-IP tunnels.
Currently it tests whether a TTL of "inherit" on a GRE tunnel actually
copies TTL of a tunnel packet from overlay, and whether changing the TTL
of a tunnel then correctly adjusts the offload.

Signed-off-by: Petr Machata <petrm@mellanox.com>
Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>