lnst.Controller.RunSummaryFormatter: filter results by result level
The format_run method will now only process Result objects with a result
level equal or lower than the result level set during the initialization
of the formatter.

Default result level of the formatter is set to IMPORTANT, but we might
change this in the future based on use experience.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>