lnst.Tests: add Iperf test module
Implements TestModule classes for the IperfClient and IperfServer that
can be used to measure network performance. Class parameters map
directly to command line arguments of the iperf3 utility, the PASS/FAIL
of the client or server jobs indicates whether or not the throughput
measurement was succesful, no check for 0 or baseline is made by the
TestModules.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>