lnst.Tests: reimplement IcmpPing as Ping
Renamed the IcmpPing testmodule as just Ping. Implements the Ping
TestModule class that can be used to test connectivity between hosts.
The module returns statistics on the measured success rate of the Ping.
The PASS/FAIL of the Ping testmodule job indicates whether or not the
measurement was succesfull, actual connectivity status needs to be
evaluated outside of this class.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>

---
v2: fixed missing if not match