lnst.Devices.Device: add ips_filter method
The method retuns a list of ip addresses that match the provided
selectors. The selectors are key=value pairs where 'key' is some
attribute of the ipaddress object, if an ipaddress object matches all
the selectors it is added to the return list of the method.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>

---
v2: fixed exception handling - "match" should be set to False