add lnst.RecipeCommon.PerfResult
The module defines classes for storing performance measurement data and
a statistics class that calculates some statistic properties of this
data for use in reporting the results.

This commit adds a prototype implementation of these classes, more work
on these is expected.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>

---
v2: fixed class names and removed a debug print