lnst.RecipeCommon: add Perf and Ping
The RecipeCommon Perf and Ping modules define BaseRecipe derived classes
(PerfTestAndEvaluate and PingTestAndEvaluate) defining common methods
used by a recipe testing performance or connectivity between two
endpoints. I expect these to be a fairly common pattern in recipes which
is why I wanted to separate them into "common" classes.

A recipe can inherit from both to combine their functionality and on
their own they don't define a fully functional recipe.

The module also defines "*Conf" classes specifying the configuration
that should be used by the Perf or Ping class. The configurations define
the endpoints that should be used and the specific configuration of the
Perf or Ping tool. This serves as an abstraction of the configuration
that is later translated to specific parameters of a test module that is
used by the PerfTestAndEvaluate and PingTestAndEvaluate classes.

The Perf module also defines a PerfMeasurementTool abstract class that
defines the interface of a class that will perform the actual perf test
using a specific measurement tool, e.g. Iperf or Netperf. The goal is to
be able to choose which measurement tool should be used.

The Ping module at this time doesn't need this because there's just one
ping module, but we could extend the code later if required.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>