lnst.Recipes.ENRT: add BaseEnrtRecipe
Adding the BaseEnrtRecipe module that defines a class with the same
name. The class inherits from both the PingTestAndEvaluate and
PerfTestAndEvaluate classes to implement a shared scheme for our ENRT
recipes. It is expected that this will go through more refactoring as we
port more recipes but the idea is that we will avoid code duplication at
all costs.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>

---
v2: define test wide (de)configuration methods to raise NotImplementedError