recipes: add netdev_cpupin alias to separate netperf and NIC pinning
Sometimes it may be useful to pin only netperf/netserver and let irqbalance
do the IRQ mangling for NIC interrupts.

Beware, that this patch changes the original behaviour. Previously, if
nperf_cpupin alias was set the irqbalance would be turned off and device
IRQs would be pinned to CPU0.

Now, if the alias is set, only netperf process would be pinned to the CPU
set in the alias value.
To disable irqbalance and pin device IRQs, you need to specify netdev_cpupin
alias with the value of CPU to be used for pinning the IRQs.

Signed-off-by: Jan Tluka <jtluka@redhat.com>