lnst.Device.Device: improve cleanup data storage
Storing cleanup data is moved into the NetTestSlave instead of the
device itself. This makes sure that the Device is returned to the state
it was in at the start of the recipe instead of at it's creation. This
has an impact on virtually created devices that can sometimes be created
with duplicated default values (e.g. name) that can conflict during
deconfiguration. But also on the generic sanity of the recipe in case
someone does some manual changes between recipes while the slave is
running, deconfiguration could have returned them back.

Finally, instead of storing cleanup variables individually, store them
in a single directory. This should also improve the implementation of
the cleanup method and make it easier to add more attributes when
necessary.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>