move lnst.Common.TestModule to lnst.Tests.BaseTestModule, add wait_for_interrupt
The BaseTestModule class should be part of the lnst.Tests package, which
gets dynamically sent to the slave when required.

I also added the wait_for_interrupt method to the BaseTestModule class.
This is functionality that we've had in the old TestModule classes but
didn't yet implement for the new Python recipes.

Even though this doesn't act as any sort of tester API or public code I
think it's a piece of code that is going to be reused a lot in many of
our test modules so it makes sense to add it to the base class.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>