lnst.Controller.MessageDispatcher: refactor wait_* methods
This commit refactors the wait_for_condition method method to improve
maintainability and readability. It also removes the unused
wait_for_job_finish method and unifies handling of machine disconnects.
Timeout handling is now part of this method instead of in the Machine
class where it doesn't make as much sense...

If a slave that isn't mapped (and is therefore not used by the currently
running recipe) the information is logged but the application doesn't
die with an exception anymore.

This includes a refactoring of the wait_for_* methods of the Machine
class, that now just delegates this functionality to the message
dispatcher.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>