lnst.Controller.Machine: expose init_connection as public method
The init_connection method is now public and the Machine object doesn't
initialize it's own connection.

Instead the connection is explicitly initialized by the SlavePoolManager
after the creation of the Machine object. This will later be used to
expose an API to the tester that will enable him to connect to slave
machines during test execution.

The method also gained a timeout parameter to limit the socket creation
duration when relevant.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>