lnst.Slave.NetTestSlave: support objects from dynamically received classes
This gives the API to instantiate objects from any dynamically received
class from the controller and then access it's attributes and methods.

For now this can only be used for RecipeCommon classes since it's the
only module (other than Devices) that the Slave can receive classes for.
But this might get extended later.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>