add lnst.Controller.SlaveObject
This is a universal class that can be used as a proxy for any
dynamically instantiated object on the slave (that was received from the
Controller).

The commit also adds the init_remote_class method to
lnst.Controller.Machine and init_class to lnst.Controller.Host to expose
this functionality to the tester.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>