add lnst.Controller.RecipeControl
Splitting off the Recipe.ctl tester API away from the
lnst.Controller.Controller class into it's own dedicated RecipeControl
class. This should be more readable and maintainable hopefully and avoid
accidentaly exposing something we didn't want to.

Currently the API provides two wait* methods, access to the Hosts
iterator and the ability to initialized a connection to a slave during
test execution. This is a new feature.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>