lnst.Controller.RunSummaryFormatter: change format for list items
In case a list item is formatted to a single line, don't add it as an
indented next line in the whole formatting but instead add it to the end
of the current line.

This should shorten the output length and improve readability for result
objects that output many short items in a list, e.g. Performance results
that include manu small samples.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>