lnst.Slave.InterfaceManager: fix deleted device handling
This commit introduces two fixes for handling situations when a Device
gets deleted. First is a small workaround for double removal of a
deleted device from the tracking dictionary. The second is an Exception
handling for calling an update on a Device object referring to a device
that is already marked as deleted. This can sometimes happen when
netlink messages get reordered so it's important to handle the
exception.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>