lnst.RecipeCommon.Perf: minor refactoring
Renaming the client and server to generator and receiver to better
illustrate the traffic flow used for performance measurement.

In case the measurement tool doesn't return measurements (either for the
generator or the receiver) the MultiRunPerf class would complain about
incompatible types (None) this was fixed by ignoring the None results.

Minor adjustment to the formating of the reporting result objects -
added information about the size of the deviation to the average value
in percentages.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>