add lnst.Recipes.ENRT.OvS_DPDK_PvP
This recipe is ported from the old
recipes/regression_tests/phase3/ovs_dpdk_pvp.xml recipe. The recipe is
also redesigned and refactored to better fit the design of our ENRT
recipes. I expect that the recipe will be further improved later but for
now this should be functional and a good base for further refactoring
when needed.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>