lnst.Devices.RemoteDevice: add caching capability
RemoteDevice objects that on the Controller act as proxies for the real
Device objects on Slaves can now be switched into a cached read only
mode.

In this mode, the RemoteDevice object stores all the property values of
the Slave Device locally on the Controller. When the tester attempts to
access a property this cached value is returned and no proxy call to the
Slave happens.

Method calls and setting new values to a property is disabled as this is
a read only mode. Unfortunatelly this means that methods that don't
change the device state (e.g. ips_filter) are also unaccessible.
However it also means that the cached RemoteDevice object can be used
even if the Device doesn't exist on the Slave anymore.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>