regression_tests: ovs-dpdk-pvp bind pings to devices
When binding ping to an IP address the kernel can choose a different
source nic if it has an IP address from the same network configured, In
our case all pings were leaving from the first (*_nic1) interface. The
same then happens on the receiver side - the replies are sent from the
default nic for that network (which is the first one again). This
resulted in the switch not getting the full information about mac
addresses that will later be used for the dpdk test.

Binding to the nic directly partially fixes this issue - the ICMP
request messages will leave from the correct nic, the replies will still
go out of the default nic for the specific network. But the switch will
get all the info it needs from the request messages.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Acked-by: Jan Tluka <jtluka@redhat.com>